backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.51.0
chromadb==0.5.7
sentence-transformers==3.1.1
pydantic==2.9.2
